refactor/1179 - moves UiController, TutorialView and TutorialStep to ui folder
