+++ src/assets/scripts/client/AppController.js
@@ -15,8 +15,8 @@ import NavigationLibrary from './navigationLibrary/NavigationLibrary';
 import ScopeModel from './scope/ScopeModel';
 import SpawnPatternCollection from './trafficGenerator/SpawnPatternCollection';
 import SpawnScheduler from './trafficGenerator/SpawnScheduler';
-import TutorialView from './tutorial/TutorialView';
-import UiController from './UiController';
+import TutorialView from './ui/TutorialView';
+import UiController from './ui/UiController';
 import { EVENT } from './constants/eventNames';
 import { SELECTORS } from './constants/selectors';
 import { speech_init } from './speech';
