+++ src/assets/scripts/client/aircraft/AircraftModel.js
@@ -15,7 +15,7 @@ import GameController, { GAME_EVENTS } from '../game/GameController';
 import ModeController from './ModeControl/ModeController';
 import Pilot from './Pilot/Pilot';
 import TimeKeeper from '../engine/TimeKeeper';
-import UiController from '../UiController';
+import UiController from '../ui/UiController';
 import {
     radians_normalize,
     angle_offset
