+++ src/assets/scripts/client/ui/TutorialStep.js
